+++ common/src/main/java/net/mehvahdjukaar/polytone/common/expressions/proxies/PlayerProxy.java
@@ -1,7 +1,9 @@
 package net.mehvahdjukaar.polytone.common.expressions.proxies;
 
+import net.mehvahdjukaar.candlelight.api.BeanGettersAliases;
 import net.minecraft.world.entity.player.Player;
 
+@BeanGettersAliases
 public class PlayerProxy extends PositionalProxy {
 
     private final Player player;